+++ xrpl4j-model/src/main/java/org/xrpl/xrpl4j/model/client/specifiers/LedgerIndexShortcut.java
@@ -5,12 +5,25 @@ import com.fasterxml.jackson.annotation.JsonValue;
 import java.util.Objects;
 
 /**
- * TODO: Javadoc
+ * A String wrapping class that defines three static constants representing shortcut values
+ * which can be used to specify a ledger index in a rippled API request.
  */
 public class LedgerIndexShortcut {
 
+  /**
+   * Request information about a rippled server's current working version of the ledger.
+   */
   public static final LedgerIndexShortcut CURRENT = new LedgerIndexShortcut("current");
+
+  /**
+   * Request information about for the most recent ledger that has been validated by consensus.
+   */
   public static final LedgerIndexShortcut VALIDATED = new LedgerIndexShortcut("validated");
+
+  /**
+   * Request information about the most recent ledger that has been closed for modifications and proposed for
+   * validation.
+   */
   public static final LedgerIndexShortcut CLOSED = new LedgerIndexShortcut("closed");
 
   @JsonValue
