+++ xrpl4j-model/src/main/java/org/xrpl/xrpl4j/model/jackson/modules/Xrpl4jModule.java
@@ -2,7 +2,7 @@ package org.xrpl.xrpl4j.model.jackson.modules;
 
 import com.fasterxml.jackson.core.Version;
 import com.fasterxml.jackson.databind.module.SimpleModule;
-import org.xrpl.xrpl4j.model.client.common.LedgerIndex;
+import org.xrpl.xrpl4j.model.client.specifiers.LedgerIndex;
 import org.xrpl.xrpl4j.model.client.specifiers.LedgerSpecifier;
 import org.xrpl.xrpl4j.model.transactions.Address;
 import org.xrpl.xrpl4j.model.transactions.CurrencyAmount;
@@ -45,11 +45,11 @@ public class Xrpl4jModule extends SimpleModule {
 
     addDeserializer(CurrencyAmount.class, new CurrencyAmountDeserializer());
 
-    addSerializer(LedgerIndex.class, new OldLedgerIndexSerializer());
-    addDeserializer(LedgerIndex.class, new OldLedgerIndexDeserializer());
+    addSerializer(org.xrpl.xrpl4j.model.client.common.LedgerIndex.class, new OldLedgerIndexSerializer());
+    addDeserializer(org.xrpl.xrpl4j.model.client.common.LedgerIndex.class, new OldLedgerIndexDeserializer());
 
-    addSerializer(org.xrpl.xrpl4j.model.client.specifiers.LedgerIndex.class, new LedgerIndexSerializer());
-    addDeserializer(org.xrpl.xrpl4j.model.client.specifiers.LedgerIndex.class, new LedgerIndexDeserializer());
+    addSerializer(LedgerIndex.class, new LedgerIndexSerializer());
+    addDeserializer(LedgerIndex.class, new LedgerIndexDeserializer());
 
     addSerializer(LedgerSpecifier.class, new LedgerSpecifierSerializer());
     addDeserializer(LedgerSpecifier.class, new LedgerSpecifierDeserializer());
